backend/homepage/requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
Django==3.1.7
django-cors-headers==3.7.0
django-filter==2.4.0
django-multiselectfield==0.1.12
django-rest-framework==0.1.0
django-summernote==0.8.11.6
djangorestframework==3.12.4
idna==2.10
Markdown==3.3.4
Pillow==8.2.0
pytz==2021.1
requests==2.25.1
sqlparse==0.4.1
urllib3==1.26.4
mysql==0.0.2
mysqlclient==2.0.3
